Draft of extracting generic line from flexible line in model
