+++ src/main/java/no/entur/uttu/error/codederror/ConstraintViolationCodedError.java
@@ -9,7 +9,7 @@ import java.util.Map;
 public class ConstraintViolationCodedError extends CodedError {
     private static final Map<String, ConstraintSubCodeEnumeration> constraintMap = Map.of(
             Constraints.FLEXIBLE_STOP_PLACE_UNIQUE_NAME,  ConstraintSubCodeEnumeration.FLEXIBLE_STOP_PLACE_UNIQUE_NAME,
-            Constraints.FLEXIBLE_LINE_UNIQUE_NAME, ConstraintSubCodeEnumeration.FLEXIBLE_LINE_UNIQUE_NAME,
+            Constraints.LINE_UNIQUE_NAME, ConstraintSubCodeEnumeration.LINE_UNIQUE_NAME,
             Constraints.JOURNEY_PATTERN_UNIQUE_NAME, ConstraintSubCodeEnumeration.JOURNEY_PATTERN_UNIQUE_NAME,
             Constraints.NETWORK_UNIQUE_NAME, ConstraintSubCodeEnumeration.NETWORK_UNIQUE_NAME,
             Constraints.SERVICE_JOURNEY_UNIQUE_NAME, ConstraintSubCodeEnumeration.SERVICE_JOURNEY_UNIQUE_NAME,
