+++ src/main/java/no/entur/uttu/error/codes/ConstraintSubCodeEnumeration.java
@@ -5,7 +5,7 @@ import no.entur.uttu.error.SubCode;
 
 public enum ConstraintSubCodeEnumeration implements SubCode {
     FLEXIBLE_STOP_PLACE_UNIQUE_NAME,
-    FLEXIBLE_LINE_UNIQUE_NAME,
+    LINE_UNIQUE_NAME,
     JOURNEY_PATTERN_UNIQUE_NAME,
     NETWORK_UNIQUE_NAME,
     SERVICE_JOURNEY_UNIQUE_NAME,
