+++ src/main/java/no/entur/uttu/model/FixedLine.java
@@ -1,0 +1,4 @@
+package no.entur.uttu.model;
+
+public class FixedLine extends Line {
+}
