+++ src/main/java/no/entur/uttu/repository/FixedLineRepository.java
@@ -1,0 +1,4 @@
+package no.entur.uttu.repository;
+
+public interface LineRepository {
+}
