+++ src/main/resources/db_init.sh
@@ -5,6 +5,7 @@ for f in ./db/migration/V1__Base_version.sql \
            ./db/migration/V3__Remove_unique_order_constraints.sql \
            ./db/migration/V4__Increased_text_column_length.sql \
            ./db/migration/V5__Increased_booking_note_text_column_length.sql \
+           ./db/migration/V6__Add_simple_line.sql \
            ./import.sql
 do
   echo "Running migration for ${f}"
